% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrient_search.R
\name{nutrient_search}
\alias{nutrient_search}
\title{Happy meal: Nutrients API Search}
\usage{
nutrient_search(app_id_nutrient, app_key_nutrient, ingr)
}
\arguments{
\item{app_id_nutrient}{Application ID for nutrient API.}

\item{app_key_nutrient}{Nutrient API key.}

\item{ingr}{The ingredient that the user wants for the nutrient information. URL-encode should be included this string.}
}
\description{
Consumers with health or diet conditions need an easy way to get the recipe of foods they want to eat during the day,
and the nutrient information of foods in their daily life.
The function allows to search for nutrient information of specified foods within several seconds,
and bring health and happiness to your daily meal.
}
\details{
The function allows to search for nutrient information of specified foods. It asks the user to insert specified queries,
including app id, app key, ingredients, etc.
Then the user can get data frame of the nutrient information matching the conditions.
}
\examples{
app_key_nutrient = Sys.getenv("nutrient_api_key")
nutrient_search(app_id_nutrient = "a81f90d1",
app_key_nutrient = app_key_nutrient, ingr = "large\%20apple")
}
\keyword{meal}
\keyword{nutrient,}
