% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipes_search.R
\name{recipes_search}
\alias{recipes_search}
\title{Happy meal: Recipe API Search}
\usage{
recipes_search(q, app_id_recipe, app_key_recipe, diet, calories, excluded)
}
\arguments{
\item{q}{The query text, what item you want the recipes focus on, for example, chicken.}

\item{app_id_recipe}{Application ID for recipe API.}

\item{app_key_recipe}{Recipe API key.}

\item{diet}{Diet label: one of “balanced”, “high-protein”, “high-fiber”, “low-fat”, “low-carb”, “low-sodium”.}

\item{calories}{Calories that each recipe serves, for example, “calories=100-300” will return
all recipes with which have between 100 and 300 kcal per serving.}

\item{excluded}{Excluding recipes with certain ingredients.  More than one food can be excluded at the same time.
Example: excluded=vinegar&excluded=pretzel，will exclude any recipes which contain vinegar or pretzels in their ingredient list.}
}
\description{
Consumers with health or diet conditions need an easy way to get the recipe of foods they want to eat during the day,
and the nutrient information of foods in their daily life. The function allows to search for recipes
matching the specified query, and bring health and happiness to your daily meal.
}
\details{
The function allows to search for recipes matching the specified query.
It asks the user to pick search for recipes, and insert specified queries, including app id, app key, diet, calories, etc.
Then the user can get data frame of the recipes matching the conditions.
}
\examples{
app_key_recipe = Sys.getenv("recipe_api_key")
recipes_search(q = "pork", app_id_recipe = "46707447",
app_key_recipe = app_key_recipe, diet = "balanced",
calories = "100-1000", excluded = "onion")
}
\keyword{meal}
\keyword{recipe,}
